+++ tests/Feature/DashboardTest.php
@@ -16,12 +16,12 @@ class DashboardTest extends TestCase
         $response->assertRedirect(route('login'));
     }
 
-    public function test_authenticated_users_are_sent_to_the_feed()
+    public function test_new_authenticated_users_are_sent_to_getting_started()
     {
         $user = User::factory()->create();
         $this->actingAs($user);
 
         $response = $this->get(route('dashboard'));
-        $response->assertRedirect(route('feed'));
+        $response->assertRedirect(route('onboarding.show'));
     }
 }